+++ regimes/regimes.go
@@ -28,6 +28,7 @@ import (
 	_ "github.com/invopop/gobl/regimes/no"
 	_ "github.com/invopop/gobl/regimes/pl"
 	_ "github.com/invopop/gobl/regimes/pt"
+	_ "github.com/invopop/gobl/regimes/ro"
 	_ "github.com/invopop/gobl/regimes/sa"
 	_ "github.com/invopop/gobl/regimes/se"
 	_ "github.com/invopop/gobl/regimes/sg"